requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary # drivers for postgres
bcrypt
pyjwt